Fix activity transitions
This fixes the bug that causes unnecessary new tasks, but also has the
effect of returning to the settings screen instead of the DeskClock if
you hit home while in clock settings and then click on clock again. If this
is not WAI, I can add in FLAG_ACTIVITY_CLEAR_WHEN_TASK_RESET when
launching the settings activity.

Bug: 7214011 Going into the clock's settings does a task transition
Change-Id: I18b28ff4ef0fb03c0f1b44db7b157fc0b24a445c