Jank improvements.
Fetching a Ringtone object causes a media player to be instantiated for each
ringtone pulled. And as it sounds, it is pretty slow which causes scrolling
in the alarm list to stutter. Fix by caching ringtone when alarm is expanded
so subsequent renders are quick.

Changed alarm updates to be asynchronous.

Bug: 7262805
Change-Id: Ie7277f2b53f3bc4a49796a41a18572f26e124dba