Fixing unique primary key error.
The undo feature is re-using primary keys. The existing check for existing keys
is not 100% correct. Fixed by actually checking the database right before an
insert. Also wrap in transaction to guarentee atomicity.

Remove undo bar when action bar is used.

Updated alarm toast to be consistent with in-place editing. Anytime alarm is
enabled, toast will be displayed.

Bug: 7299177
Change-Id: Iedce94b2ddf06002c2da239af9299151d3e267ab