Read cities names for DB and not from list
Bug: 7413861

The list of user selected cities in world clock was using the name from the usr's list and
not from the DB. When a change of locale is done on the device, the cities name were not changed.
Changed the WorldClock adapter to use the DB name and timezone if available.

Change-Id: I11b44c61c792ce0821f8a8e34f9c9086d6bfb66f