Add intents in the broadcast listener of the widget
bug: 7511464
Added intents to listen to TIME_TICK, time changes and screen on in order to update the
next alarm info and the city's day in the alarm clock.
Listening to to TIME_TICK is not needed for the clock but it is needed to update
the day in the world clock when the time passes midnight in a selected city.

Change-Id: Id07a4a0d6722adb9c896b1f7f1c71c594e500523