Use trilinear filtering to draw the analog clock
The clock is often drawn scaled down from its original size. This
change turns on trilinear filtering which makes the resampled
result look a little nicer. The difference is visible even on
xhdpi devices.

This change relies on a new XML attribute introduced in API level
18 so the makefile was updated to use the current SDK. The attribute
will be safely ignored on previous versions of the platform.

Change-Id: I506e13398496b66b5dd8f2d789ea338d99371635