Allow tablets to freely rotate the alarm alert activity.
Tablets were switching to their defualt orientation due to the
nosensor orientation set in the manifest, and were then being changed
to unspecified once the activity was created. Doing it the opposite,
so phones and tablets start as unspecified and only phones are
then designated as "nosensor" during onCreate() allows us to get
the desired behavior.

Bug: 8728671
Change-Id: I29ac593ccbb65953fa35dad38110accc9c531594