Configure HandleApiCalls launchMode to singleTask
Bug: 18436499

It makes sense for the HandleApiCalls activity to be a singleTask since
it always finishes itself right after starting the DeskClock activity,
which is also a singleTask. This also fixes the bug when HandleApiCalls
is not triggered at all on subsequent intents due to "Warning: Activity
not started, its current task has been brought to the front".

Change-Id: I29f4b8e2caef1f48b988886cb8596edfb2384520