Remove the prior state preference when deleting a TimerObj
TimerObj can read/write/delete itself to SharedPreferences.
A field named "timer_prior_state_" was added to the read/write
methods but was missing from the delete method. This makes it
possible to delete a timer then create a new timer that may
begin with a random "prior state".

This fix is to actually delete this field in addition to all
others recorded for a TimerObj in SharedPreferences.

Change-Id: I484b19bbdbbc7367bcd2880a865c7509c4859a6c