Refactoring HandleVoiceApiCalls
Refactored HandleVoiceApiCalls to call private classes for AsyncTask.
Changed the default behavior of DELETE_CLOCK when a city is not
specified. Now it just opens the CitiesActivity where a user can choose
which city they want to delete.
Refactored notifySuccess and notifyFailure methods to make them static
so that they are called from other classes and work for Async methods.
Also added these two methods for every outcome of any voice action.
Refactored HandleVoiceApiCalls to use the App context in all the async
methods so that it doesn't try to use the activity context after the
activity was finished.
Changed DISMISS_UPCOMING_ALARM to use SEARCH_MODE's

Change-Id: I897617659523f45a6d8d6b06f6469f0e3cb421da