Change the toggle to off when DISMISS'ing 1 time alarms
When DISMISS_ALARM is triggered right now there is
absolutely no indication that an alarm is proactively dismissed.
It'll be even more of a problem when we change the Dismiss
window from 2 to 24 hours.
This CL sets the toggle to 'off' for 1 time
alarms so that it's clear that the alarm is dismissed.

Change-Id: I6d0fd0cafd5b3af8200767579a1a6dd46005df26