Handle timer fab button when timer state is STATE_DELETED.
Bug: 21391070

In earlier versions of Clock, Timers are set to STATE_DELETED when
stopped post-firing. In version 4.0.1+, timers are marked STATE_DELETED
when they are actually deleted.

This discrepancy causes the timer fab button not to be refreshed when
the on-screen timer is from an older version and in state STATE_DELETED.

Change-Id: I280a9befae623c501f274ffe99dd6f2628cdefee