Change digital clock widget font to match new lockscreen font.
Bug: 21615903

The lockscreen font for clock has changed in M+. For a unified user
experience, change the widget style for API23+ to use the same fonts.

With that comes a more spaced-out font for the date/next alarm,
necessitating a decrease in that font size for some devices.

In the course of this, discovered that the textclock in ClockFragment
got cut off when Display is set to huge font. Also adjusted the font
size there.

Change-Id: Ic1fc46097bc00a0e8957c0db508f470ae0585061