Add provisions for translators to manually specify city name indeces.
Bug: 21790880

Allow translation the option to provide the index that a city should be
sorted under. This is to circumvent the indexing problems on platforms
that used ICUv53 (L+).

By default, continue using the first letter of the city name as the
index. Translation may specify a different index by modifying the
[index] value in the following string:
<item>[index]<xliff:g id="separator">=</xliff:g>[city name]</item>

This means that indeces can now consist of multiple characters, where
before they were defaulted to a single character.

Change-Id: I0a437bdf941140be1c4052d212ccdd2195a3daec