Add backup/restore support
We use the new fullContentBackup feature introduced in M. The
wrinkle that exists in our case is that the preference file
could be either:

com.android.deskclock_preferences.xml
or
com.google.android.deskclock_preferences.xml

depending on whether DeskClockGoogle overlayed DeskClock or
came as part of the device image.

So, a BackupAgent is also configured and used to intercept the
writing of the preference file on the restore device. The
preference filename is then altered to include the package name
of the receiving application, guaranteeing that the preferences
(which store world clocks, timers and stopwatch data) survive
intact.

Bug: 21818176
Change-Id: Ie4652ef8375b8db4ab6e2ae3dc40820dae618ddf