Prompt for READ_EXTERNAL_STORAGE perm on custom ringtone selection
If a ringtone stored on external storage (e.g. sd card) is selected
and the READ_EXTERNAL_STORAGE permission has not yet been granted
by the user, prompt the user to grant it. This prompt will be
displayed after the ringtone chooser is hidden if the selected
ringtone is external.

Bug: 20273223
Change-Id: I1c52aaaf3b36a279202c3d210a4f204637f43e53