Add custom ringtones for timers.
Bug: 18366475

Add an option in the SettingsActivity to change timer ringtone.
In order to keep currently satisfied users happy, default to using
the current timer_expire.ogg ringtone. If the user opens the
timer ringtone preference, they'll be able to pick from
RingtoneManager's alarm selection. This also means that they will
never be able to get timer_expire back unless they clear their
SharedPreferences, because timer_expire is not one of
RingtoneManager's provided ringtones.

Before the user selects a RingtoneManager ringtone, leave the
summary for the timer ringtone preference blank so the user will
have less of an impression that timer_expire is a selectable option.

Change-Id: I513514d9add3e67e22d539d54eeea236677e5874