Create the default alarms with vibrate on by default
Many users simply alter the sample alarms in the Clock app
to create their first functional alarm. Those sample alarms
are now created with the vibrate feature enabled. It is
better that users are introduced to the vibrate option and
then learn how to disable it than to never learn about it at all!

Change-Id: Ifee3c7d7fdeb7cad2f796d172a83813d668d2a0e