Refactor Cities (World Clocks) to use new unified DataModel
This refactoring replaces CityObj with City, an immutable description
of a city and its associated clock information. City objects are fetched
via a DataModel class that is the single point of reference for all City
data. This allows the data to be shared across the entire application.
In particular, DigitalWidgetViewsFactory, ClockFragment and
CitySelectionActivity all draw upon the same model to determine what
cities exist, which ones are selected, and what the preferred display
order shall be.

Change-Id: I7c2eaa31b596b5333213415af58a7bb46ebb370e