remove requests for READ_EXTERNAL_STORAGE permission
Very late in the M release, a change was made to no longer
require the READ_EXTERNAL_STORAGE permission to view
the titles of external ring tones. Consequently, we no longer
have a reason to acquire this permission in the application.

As well, the storage of the ringtone to play for expired timers
has been moved from Utils into the DataModel.

Bug: 22919062
Change-Id: I9209cd3e6efaee1063c1c04bca15b0ea61ffc4d8