Present Clock's timer beep as the default ringtone
A "settings" package was introduced as a byproduct of this CL.

When choosing the preferred ringtone for timers, the default
selection should play timer_expired.ogg which ships with the
application. This is the only sound played on all prior version of
the application.

Known Issue: the dialog will label the default as
"Default alarm sound". This will be fixed in a later CL.

Change-Id: I4e9c6663c597038ec273b699cdda3a036b376350