Refactor Stopwatch to use new unified DataModel
This refactoring introduces Stopwatch and Lap domain objects.
Instances of these domain objects are fetched via a DataModel
class that is the single point of reference for all stopwatch data.
This allows the data to be shared across the entire application.
In particular, the stopwatch notification and StopwatchFragment
all draw upon the same model to determine what the state of the
stopwatch is and what laps have been recorded.

Change-Id: I6bda1fda73568cb312b167f1f6b10bd1abc99d62