Avoid displaying two copies of the same alarm
Bug: 17903090

When an alarm is missed the alarm_instances table actually contains two
entries for the same alarm.

- one represents the missed alarm
- one represents the next upcoming alarm

The query that joins alarm_templates to alarm_instances would return
two rows for the same alarm in this case. Since the alarm_instance data
is used offer a predismiss button in the UI, it is fine to ignore the
missed alarm in this join.

Change-Id: I720e48c9274b42d3cda9d57d3eef099c5dcd3902