Simplify HandleDeskClockApiCalls, TimerService and StopwatchService
HandleDeskClockApiCalls used to pass through all timer and stopwatch
mutations to the corresponding Service to complete. This complicated
the logging of the corresponding event unnecessarily.

After this change:

- StopwatchService is only invoked from stopwatch notifications and may
now assume the event label

- TimerService is only invoked from timer notifications and AlarmManager
and may now assume the event label

- HandleDeskClockApiCalls.EXTRA_FROM_NOTIFICATION has been replaced
with HandleDeskClockApiCalls.EXTRA_EVENT_LABEL which explicitly
identifies the label of the event

Change-Id: Ic9796c43e826f82138ee431d3abb02524c0146ba