Fix DeskClock recreation after altering settings
Bug: 25121359

Calling Activity.recreate() while the activity is paused will create
a new copy of the activity that remains paused, even if it is the
foreground activity.

To work around this, we delay recreating the activity until after
it identifies as being resumed. This produces a noticeable flash
but is far better than the alternative...

Change-Id: Iae770a25fb5b6f80d50a321a0f2f4cd792eceb10