Avoid jank in the tabs from a cold open
Bug: 24977472

By default, the menu is inflated after the first draw. A proactive
call to Toolbar.inflate(...) during creation defeats this default
behavior.

A second problem is that the menu that is inflated by default for
DeskClock contains the search menu item which is briefly shown.
By defaulting that menu item to be invisible and only making it
visible when its corresponding controller exists, the
dancing/flashing effect is avoided.

Change-Id: I933ab9a59b471abe26e1cf07385e4a4c128d1808