Fix fab disappearing on rotation when no timers exist
Bug: 25779786

The issue here is that TimerSetupView actually customizes the fab.
These customizations include setting it invisible. It was a poor design
to share the fab with this view. So, the fix is that the view now holds
its own fab-like button that it is free to manipulate without fear of
altering the app-wide fab state unintentionally.

Change-Id: Ie83b03c98db4f77e0be912b122b6f74c2c7e6c8a