Do not attempt to display deleted timer
TimerIntentsTest actually caught this bug. The intent to show
the timer tab requested that a freshly deleted timer be displayed.
This can produce exceptions if it was the last timer.

When a timer is deleted, the timer tab can be shown but no
specific timer should be requested for display.

Change-Id: I7024fbe5a154d5ecb8d563c80c09044d3917ab50