Attempt to avoid early death of AlarmService
Bug: 26058053

Evidence shows that AlarmService is being destroyed a couple 100
milliseconds after starting. This can be catastrophic as it typically
means that the ringtone playback is ended before it even sounds.

This CL balances the binding/unbinding calls in AlarmActivity to
AlarmService to encourage the UI to stay in the foreground and
keep the service alive and the ringtone playing.

This CL also removes unnecessary cancel()ations of notifications.

Change-Id: I59b35dc8404dbd1f8d74285aa34a7cd81ec5545b