TimerSetupView now repurposes fab and buttons
TimerSetupView used to have its own fab and a custom button for
the X icon that cancels timer creation. It now utilizes the application's
fab and buttons. This ensures that fab/button locations and
animations are uniform.

Child fragments and views no longer have direct access to the fab
and buttons outside of specific callbacks. This design attempts
to remove past bugs where irrelevant views altered the state of the
fab at inappropriate times.

Change-Id: If8c69246e915f6357df80a6d8107f7b2b5a6172e