Support a variety of am/pm string sizes when formatting times
Bug: 26015206
Bug: 26014517

ag/786215 introduced a small regression which assumed that the
meridiem (am/pm) portion of a time string should always be formatted
at the same proportional size as the remainder of the time display. In
truth, the user interface contains multiple ratios which were
inadvertently coalesced to a single ratio.

For example, the alarms tab formats alarm times with smaller am/pm
strings than the rest of the application. The display of selected world
clocks has been changed to share the same ratio with the alarms tab
in this change.

This change also enables the meridiem display on the digital widget.

Change-Id: I351c816fd522aa32c445b85ac19f619e0a030a43