Remove special characters from comparison when searching cities
Bug: 26346964

All of ' ', '-', '.' and ''' are now removed to make prefix matching more
liberal and the job of locating the correct city easier. For example,
"St. Lucia" is now matched by any of these search terms:

- StLucia
- St Lucia
- St.Lucia
- St. Lucia

I experimented with contains() logic, but it doesn't feel right and breaks
user expectation in many ways, so startsWith() was left in place.

Change-Id: I0b92bbdef677c7309cc551eefbcee8c5e7376263