Avoid IllegalAccessError while displaying missed alarm
Bug: 26714950

AlarmStateManager attempted to adjust the selected tab to be
the alarms tab in advance of firing an Intent that would scroll
to a given alarm. Since this occurred on a background thread,
the sanity thread check in UiDataModel was tripped.

The good news is that the receiving logic in AlarmClockFragment
*already* switches to the alarms tab when
AlarmClockFragment.SCROLL_TO_ALARM_INTENT_EXTRA is
present in the Intent, so the problematic call is unnecessary and
was removed.

Change-Id: I742fa92ce7e75b9064436f09dfcce3164fd6bc65