Do not assume current stopwatch lap has positive duration
Bug: 27171498

After a device reboot the realtime clock is reset. Because of this, the
app is notified of the BOOT_COMPLETE and resets the stopwatch.
There exists about a 10 second window of opportunity before the
BOOT_COMPLETE broadcast in which the app may be opened and
the realtime clock information may be used to compute stopwatch
laptimes which would be negative. This case is now handled gracefully
rather than crashing on a sanity check.

Change-Id: I0eb4da0a9a55b5abc8d6abaa5a2e046821d39a2e