Stop using custom fonts
Bug: 27145641

There was a time when Roboto was not a monospaced font. For
usage such as stopwatch and timers, this was problematic as the
display appears to jitter around. The solution at the time was to use
a custom font that *was* monospaced.

It is no longer necessary and we return to using the default sans-serif
font (which is Roboto... for now) for consistency throughout the UI.

Change-Id: I61e8052255b44569000fa33e564d900b5f17dd93