Overhaul the digital clock widget
Bug: 7346144
Bug: 26747882
Bug: 27391440
Bug: 27226990
Bug: 27282291
Bug: 18410094
Bug: 22124815
Bug: 27459689
Bug: 24003653
Bug: 27118501

The digital clock widget now scales to the size it is granted by the user.
The scaling is accomplished by dynamically choosing a font size, and
testing (through offscreen measurement) if the chosen font will
successfully fit in the widget's allocated space. This technique should
be successful across all locales, fonts and sizes.

Note, as of this CL, digital clock widget no longer contains world cities.
It only displays the current time, date, and next alarm time. Plans exist
to create a new stand-alone widget for just the world cities.

Change-Id: I2890f7a1aa8687218f1bbef485dc1154c65bde15