Stop using dedicated view to indicate silent state of alarm
Bug: 23940484

Before this change a line existed in expanded and collapsed alarms
that would indicate "Silent ringtone" or "Alarm volume muted" for each
alarm. This has been removed in favor of a new design.

After this change, if the user selects a non-silent ringtone, a filled bell
icon is displayed next to the name of the ringtone in expanded alarm
cards. If the user selects the silent ringtone, an empty bell icon is
displayed next to the localized word "Silent".

A separate solution for incorporating "Alarm volume muted" will be
proposed in the future.

Change-Id: Ia4abdeb18434ab8fd8a7b45d2c49ead25ed6eab1