Fix the alarm list animations
Bug: 25421235

Specifically, after this change these animations are improved:

- create alarm
- delete alarm
- enable/disable alarm

When creating, deleting or mutating alarms the ClockProvider
is consulted discretely many times and produces many notifications.
This in turn causes many rapid reloads of the alarm data and
interrupts ongoing animations in the alarm list. To sidestep this
problem for the time being (until a larger refactoring can remove
the root problem of spammy provider notifications), updates to the
alarm RecyclerView are queued and processed only after the ongoing
animation completes.

Change-Id: Iec88fa944a498516424ca65744a11e46fe42d39d