Avoid IllegalStateException while creating new alarm
Bug: 27724302

ag/887398 added the ability for AlarmClockFragment to delay
updates to RecyclerView until current animations completed.
It is necessary to also delay related work (expanding the correct
alarm, scrolling to new alarms) that will change the layout of the
RecyclerView lest we suffer the exception mentioned in the bug.

This CL also restores the behavior that scrolls to and expands
new alarms.

Change-Id: Ic010b50ea20f081f3e5d7a0d8146148d4026afb4