Show toast when DND is blocking alarms
Bug: 17505626

This is purely informational to help set the user's expectation about
whether alarms will fire or not.

Also note that it reports the message "Device is set to total silence"
based on the *current* DND setting. No calculation is done to predict
the DND state at the time of the alarm (e.g. total silence is set to
expired before the alarm fires).

Change-Id: I1fa59391154d81f88a755f4c4e47d9de10c841f3