Fade out stopwatch laps that pass under fab buttons
Bug: 28842209

The effect when lap times pass behind the fab (or the Share button
in landscape) is poor. Fading those bottom lap times increases the
contrast and is easier to decipher.

This CL also moves the background color computation out of Utils
and into ColorModel, a delegate of UiDataModel, which internally
handles updating the color each hour. These color changes are made
available to the activity and its fragments via a new
onWindowBackgroundChanged(...) callback

Change-Id: I8ab0efa849c0e0e33b6f8ff4d8e66bdd70d090f1