Stopwatches and timers continue to work after device reboot.
Bug: 8101723

Upon device shutdown and reboot, the stopwatch will resume state. If
the stopwatch was running, it will continue to run as well as account
for the time that the device was off. Additionally, all timers that
are on the device will resume state. If the timer was running, it will
continue to run on device reboot. If the timer was missed by less than
1 minute, the timer will still go off. If the timer was missed by more
than 1 minute, the timer will go to a new "missed" state and give the
user a notification that a timer has been missed.

Change-Id: Ib7f199ef5be4714b831450933c574f39824d15dc

Stopwatches and timers continue to work after device reboot.

Bug: 8101723

Upon device shutdown and reboot, the stopwatch will resume state. If
the stopwatch was running, it will continue to run as well as account for
the time that the device was off. Additionally, all timers that are on the
device will resume state. If the timer was running, it will continue to run
on device reboot. If the timer was missed by less than 1 minute, the
timer will still go off. If the timer was missed by more than 1 minute, the
timer will go to a new "missed" state and give the user a notification
that a timer has been missed.

Change-Id: I831d7a5172a3cede595db1b62e503ecc8c4c55c2