Scrollbar no longer touches analog world clocks.
Bug: 29393258

Right gutter for world clocks is now applied inside of the world clock
items versus in the clock fragment itself. This allows for the scroll
bar to appear all the way to the right of the screen which moves it
out of the way of the scroll bar.

When in portrait, left gutter is also applied inside of the world clock
items so that the overscroll looks balanced.

Scrollbars now reach the bottom of the screen when you scroll to the
bottom of the list.

Change-Id: Id3917727f38dd21cd2ce340d17347090f40269d6