Hide drop shadow when stopwatch is reset and alarms list is empty
Bug: 32644925
Bug: 32704450

In landscape, it is possible to record enough stopwatch laps that
the drop shadow is displayed. Upon resetting the stopwatch, the
laps are cleared and the laps list RecyclerView is set to GONE, but
that change in visibility goes unnoticed.

Android does not grant a clean way to observe that visibility change
so the drop shadow state is updated by the data observer.

Similarly, the alarms RecyclerView is set to GONE when the last
alarm is removed. The drop shadow state is now updated if the
adapter is emptied.

Change-Id: If4e9cb8856eefbf578a218d04a66b99cc66f562c