Fix the fill area when transitioning from expanded to collapsed
Bug: 32203879
Bug: 32703777

Expanded alarms have a drop shadow that is implemented as a
background image and thus doesn't count as content area.
Collapsed alarms have no drop shadow. When animating between
them this must be accounted for, lest the touch fill color not
bleed to the bottom edge because of the vestigial expanded
background drop shadow image.

Change-Id: I7cc91c2532477e7f28f4ac89ac8bfd4888f76cee