Stop misrepresenting alarm time on DST days
Bug: 27654884

The alarms list displays the alarm time by formatting a Calendar object.
The Calendar object is always created on the current date. If the
current date happens to be the day DST is applied and the alarm time
falls in the missing hour (in the spring), Calendar adjusts the hour ahead
an hour, causing the alarm time to be incorrectly displayed on screen.

The Calendar object is now created on a fixed non-DST day to avoid
applying DST rules inappropriately.

Change-Id: Ibf3f86a84171edcf819a132fc3f3efdd4f32dd5d