Switching between LTR and RTL preserves the selected tab
Bug: 32704236
Test: Open app, note the selected tab, adjust the language to one
layed out in the opposite direction, reopen app, check selected tab.

UiDataModel has always stored the selected tab as an enumerated
value independent of position within the UI. This same enumerated
value is now attached to ViewPager fragments and Tabs in TabLayout
to simplify the task of matching the selected tab enum with the
appropriate counterpart within the UI components.

In addition, the ViewPager no longer saves and restores the selected
fragment. That is accomplished by application code rather than the
component itself. Also, ViewPager fragments are now registered with
a tag that is position-independent, which means they can be easily
reused if the app is rebuilt after a locale change.

Change-Id: Ib3ed2490982f1adefb847cc9647e671bed891a0c