Animate fab and timer view change in parallel
Bug: 32703589

These animations used to occur serially. To parallelize them,
new UpdateTypes were added to only shrink and only expand
the fab and buttons. This allows these animations to be
synchronized with other animations that alter the view.

Also, creating the new timer had to be delayed until after
hiding the fab and buttons so that their state was consistent
as they shrunk. This mirrors the delayed removal of timers
that already existed.

Change-Id: I381b86fd55d5d51c522138c2b51996594da7f85e