Avoid setting default home timezone to an illegal value
Bug: 33418631
Test: Flash device. Open Clock. Verify the home timezone is absent.

Adjusting the default home timezone via Settings presents a subset
of all available TimeZones. After a fresh install of Android, it is possible
for Clock to launch before setup wizard completes and the timezone
of the device becomes known. In these cases, attempts to write an
illegal TimeZone (absent from the set of TimeZones to choose from in
Settings) may occur. These attempts are now ignored.

Change-Id: I0e9db67f2fc3250487d11899a36aa867f894a0b5