Handle SecurityExceptions while releasing file permissions
Bug: 33461345
Test: Manual test of RingtonePickerActivity

When custom ringtones are created, a persistable file permission to
read (play) the ringtone is taken. When the ringtone is deleted, the
permission is released. If the file was removed from the file system
prior to releasing the permission, a SecurityException occurs because
the permission in question is no longer held by the application. This
SecurityException is now handled gracefully.

Change-Id: Ia6fe9dbe2d1129bd6c081ae081da52fb7e61e334