User is now informed when a custom ringtone in unplayable.
Bug: 34132184

When a user moves, renames, or deletes a file which was previously
selected as a custom ringtone, the clock is unable to play it back.
This is now conveyed to the user as a warning badge on the ringtone
selection screen. The user will now be prompted to remove the
ringtone if they tap on the badged ringtone.

Change-Id: I913b0c194dc463ff51b59b7c8e600ab18507b82d
(cherry picked from commit 84994eb7985da1cdfa379b73bdee31ab548c102d)